ci: fix all build warnings + infra fixes (kmod, IPFS HTTP API, KCFLAGS)
* ci: fix Dockerfile (add kmod), suppress zero-length/indent/psabi warnings, use IPFS HTTP API

- Add kmod to Dockerfile so depmod runs during modules_install
- Add -Wno-zero-length-bounds (90 warnings), -Wno-misleading-indentation (4),
  -Wno-psabi (20 GCC 9.1 ABI notes) to KCFLAGS in both workflows
- Replace ipfs CLI with kubo HTTP API (127.0.0.1:5001) in release.yml
  since github-runner has no local IPFS repo (~/.ipfs)

* kconfig: remove !DISP2_SUNXI guard from FB_CMDLINE depends

DRM selects FB_CMDLINE unconditionally, which triggers 4 unmet-dependency
warnings when DISP2_SUNXI is also enabled. The guard was ineffective
(select overrides depends) and both DRM (for PowerVR GPU) and DISP2_SUNXI
(for display pipeline) must remain enabled on TSP.

* xr829: fix sprintf overlap, format strings, return type, and IS_ERR_VALUE warnings

- fw_dbg_inf.c: replace sprintf(buf,"%s...",buf,...) UB with
  len+=sprintf(buf+len,...) across 10 debug-dump functions (39 -Wrestrict)
- queue.c: %d -> %zu for size_t fields (10 -Wformat)
- wsm.c: %d -> %u for u32 buf_len (1 -Wformat)
- etf.c: add missing %s specifier for __func__ (3 -Wformat-extra-args),
  add return 0 to xradio_etfcli_data_init (1 -Wreturn-type)
- platform.c: cast to (long) before IS_ERR_VALUE (1 -Wint-to-pointer-cast)

* bsp: fix unused variable/function, sprintf overlap, IS_ERR_VALUE, and array-bounds warnings

cmsis-dap:
- DAP_config.h: mark 4 static functions __maybe_unused
- DAP.h: mark 2 static functions __maybe_unused, add #ifndef guards
  for DAP_SWD/DAP_JTAG macros (2 -Wundef)
- SWD_main.c: remove 7 unused variables
- swd_host.c: remove unused variable, fix declaration-after-statement

analog.c: use temp buffer + strlcat for snprintf overlap (-Wrestrict)
phy_device.c: use len+=sprintf pattern, add GENPHY_DRV_RTL8211F enum
  entry for excess-elements warning
xradio_btlpm.c, sunxi-wlan.c: cast to (long) before IS_ERR_VALUE
cpuinfo.c: remove unused serial variable
sunxi_timer_alarm.c: remove 2 unused flags variables
gc02m1_mipi.c: pragma-suppress false-positive -Warray-bounds

+++ drivers/misc/sunxi-rf/sunxi-wlan.c
@@ -137,7 +137,7 @@ int sunxi_wlan_get_oob_irq(void)
 	pdev = wlan_data->pdev;
 
 	host_oob_irq = gpio_to_irq(wlan_data->gpio_wlan_hostwake);
-	if (IS_ERR_VALUE(host_oob_irq))
+	if (IS_ERR_VALUE((long)host_oob_irq))
 		dev_err(&pdev->dev, "map gpio [%d] to virq failed, errno = %d\n",
 			wlan_data->gpio_wlan_hostwake, host_oob_irq);
